requirements.txt
elevenlabs==0.2.26
python-dotenv==1.0.0
requests==2.31.0 